q ability and rendering is functional in game, might be balanced differently in the future

+++ entities/Entity.h
@@ -33,7 +33,7 @@ public:
     /// The total animation frames (columns) in the sprite sheet
     int totalFrames = 1;
     /// Which animation (row of frames) is playing. If negative, don't render the Entity.
-    int animation = 0;
+    float animation = 0;
     /// The how many frames per second the animation plays at.
     float speed = 1;
     /// The current frame (column) that the animation is on
